% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{addClustering}
\alias{addClustering}
\alias{addClustering,SCSRDataModel-method}
\title{addClustering}
\usage{
\S4method{addClustering}{SCSRDataModel}(obj, cluster.id, coordinates = NULL, method = NULL, plot = TRUE)
}
\arguments{
\item{obj}{an object of class SCSRDataModel}

\item{cluster.id}{a dataframe with cluster id and names}

\item{coordinates}{a dataframe with x and y coordinates for the projection}

\item{method}{a character, giving the projection method}

\item{plot}{a logical}
}
\value{
A SCSRDataModel with cluster definition
}
\description{
Add the cell clusters, id, names and 2D coordinates for graphs
}
\details{
The ` obj` argument must be a a SCSRDataModel object, obtained 
by applying the ` dataPrepare()` function to the counts matrix (either raw 
or already normalized).

The ` cluster.id` argument must be a dataframe containing 2 
columns: cluster  ids and names. It must have as many rows as there are 
columns in the matrix counts.

The ` coordinates` argument must be a dataframe with 2 columns 
containing x and y coordinates in the 2D projection. It must have as many 
rows as there are columns in the matrix counts.

The ` method` defines the type of 2D projection (umap, pca, 
tsne, ...).

The ` plot` argument must be a logical to chose whether to plot 
the projection

The ` coordinates` and ` method` arguments are not mandatory but 
are needed in order to do cell visualization.
}
\examples{
message('addClustering')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--add Clustering')
obj <- addClustering(obj,cluster.id = sample(1:5,size = 20,replace = TRUE))
}
