% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{cell.representation<-,SCSRDataModel-method}
\alias{cell.representation<-,SCSRDataModel-method}
\title{cell.representation setter (internal use only)}
\usage{
\S4method{cell.representation}{SCSRDataModel}(x) <- value
}
\arguments{
\item{x}{A SCSRDataModel object}

\item{value}{valut to be set for SCSRDataModel}
}
\description{
cell.representation setter (internal use only)
}
\keyword{internal}
