% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{cell.representation}
\alias{cell.representation}
\alias{cell.representation,SCSRDataModel-method}
\title{Model cell.representation accessor}
\usage{
\S4method{cell.representation}{SCSRDataModel}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
Model cell.representation accessor
}
\keyword{internal}
