% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_classifier.R
\name{cell_classifier}
\alias{cell_classifier}
\title{Cell classifier}
\usage{
cell_classifier(
  data,
  genes,
  markers = markers_default,
  tsne = NULL,
  plot.details = FALSE,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or
UMI counts (note : rownames(data)=genes)}

\item{genes}{a character vector of HUGO official gene symbols of length n}

\item{markers}{a data frame of cell type signature genes}

\item{tsne}{(optional) a table of n rows and 2 columns with t-SNE
projection coordinates for each cell}

\item{plot.details}{a logical (if TRUE, then plots the number of cells
attributed to one cell type, see below)}

\item{write}{a logical}

\item{verbose}{a logical}
}
\value{
The function returns a list containing the thresholded table, the
maximum table, the raw table, a cluster vector, a pheatmap object
(see pheatmap library) and the cluster names. The maximum table is a special
thresholded table where in every column only the maximum gene signature is
kept. It can be used to force the classification of every cell. The pheatmap
object contains the thresholded matrix, which is the one to be used for most
applications, in a heatmap representation for immediate display purpose.
}
\description{
Classifies cells using cell type specific markers.
}
\examples{
data = matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) = paste("gene",seq_len(50))
markers = matrix(paste("gene",seq_len(10)),ncol=5,nrow=2)
colnames(markers)=paste("type",seq_len(5))
cell_classifier(data,rownames(data),markers)
}
