% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{cellClassifying}
\alias{cellClassifying}
\alias{cellClassifying,SCSRDataModel-method}
\title{Cell classifier}
\usage{
\S4method{cellClassifying}{SCSRDataModel}(
  obj,
  markers = markers_default,
  projection.method = "tsne",
  most.variables = TRUE,
  plot = TRUE,
  write = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{an object of type SCSRDataModel}

\item{markers}{a data frame of cell type signature genes}

\item{projection.method}{'tsne'}

\item{most.variables}{a logical}

\item{plot}{a logical (if TRUE, then plots the number of cells
attributed to one cell type, see below)}

\item{write}{a logical}

\item{verbose}{a logical}
}
\value{
A SCSRDataModel with cluster definition
}
\description{
Classifies each cell using cell type specific markers.
}
\details{
The ` markers` argument must be a table with cell type gene
signatures, one cell type in each column. The column names are the 
names of the cell types.

The *markers.default* table provides an example of this format.

If no projection is provided in the object, then the function 
will not display the cells.

If ` plot` is TRUE, then the function plots the number of cells
attributed to a single cell type as a function of the threshold applied to
the normalized gene signature average.

If `most.variables` is TRUE, then the function uses the most
variable genes matrix counts if it exists in the object.

If ` write` is TRUE, then the function writes four different
text files. (1) The 'raw classification matrix' provides the normalized
average gene signature for each cell type in each individual cell, a number
between 0 and 1. This matrix has one row per cell type and one column per
cell, and the sum per column is 1. Row names are the cell type names (column
names of the markers table) and the column names are the individual cell
identifiers (column names of `data`). (2) The 'thresholded classification
matrix', which is obtained by eliminating all the values of the 'raw
classification matrix' that are below a threshold a\*. In practice, a\* is
automatically determined by the function to maximize the number of cells
that are assigned to a single cell type and all the cells (columns) assigned
to 0 or >1 cell types are discarded. The number of cells assigned to a
single type depending on a\* can be plotted by using the parameter
`plot=TRUE`. (3) A cluster vector assigning each cell to a cell type. Note
that a supplementary, virtual cluster is created to collect all the cells
assigned to 0 or >1 types. This virtual cluster is named 'undefined'.
(4) A table associating each cell type to a cluster number 
in the cluster vector.
}
\examples{
message('cellClassifying')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--cell Classifying')
markers <- matrix(paste('gene',seq_len(50)),ncol=5,nrow=10)
colnames(markers) <- paste('type',seq_len(5))
obj <- cellClassifying(obj, markers = markers)
}
