% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{cellNetHeatmap}
\alias{cellNetHeatmap}
\title{Heatmap overview of cellular networks}
\usage{
cellNetHeatmap(
  obj,
  selected.populations = NULL,
  genes.to.count = NULL,
  only.R.in.genes = TRUE,
  use.proportions = FALSE,
  low.color = "white",
  high.color = "royalblue3",
  thres = NULL,
  col.fun = NULL,
  row.bar.color = "darkcyan",
  col.bar.color = "seagreen",
  pop.font.size = 10,
  title.font.size = 12,
  legend.font.size = 10,
  bar.plot.height = NULL,
  rect.border.width = 4,
  bar.num = FALSE,
  bar.num.font.size = 8,
  original.order = FALSE
)
}
\arguments{
\item{obj}{A SCSRNet or SCSRNoNet object.}

\item{selected.populations}{A vector of cell population names
to consider in the plot. By default, all the populations are
considered.}

\item{genes.to.count}{A vector of gene names for counting or enrichment
analysis.}

\item{only.R.in.genes}{A logical indicating whether \code{genes.to.count}
should be regarded as containing receptor gene names only.}

\item{use.proportions}{A logical to choose between representing the
proportion of genes in \code{genes.to.count} or its enrichment.}

\item{low.color}{The color to be used for the lowest value.}

\item{high.color}{The color to be used for the highest value.}

\item{thres}{A higher threshold imposed to the values.}

\item{col.fun}{A function that returns a color based on the values in the
matrix. When no such function is provided, a function is generated from
\code{low.color} and \code{high.color} with a linear gradient.}

\item{row.bar.color}{The color of the barplot counting the total number of
received LR interactions.}

\item{col.bar.color}{The color of the barplot counting the total number of
emitted LR interactions.}

\item{pop.font.size}{Font size for cell population names.}

\item{title.font.size}{Font size for row and column titles.}

\item{legend.font.size}{Font size for the legends.}

\item{bar.plot.height}{Heigh (or width) or the two barplots.}

\item{rect.border.width}{White border width around the colored
rectangles of the heatmap.}

\item{bar.num}{A logical indicating whether numbers should be written on top
of the barplots.}

\item{bar.num.font.size}{Font size of the totals on the barplots.}

\item{original.order}{A logical indicating whether matrix rows and columns
should be left in their original order. If not (the default), they are
reordered by computing two dendrograms that are not displayed.}
}
\value{
A heatmap is displayed reflecting population pairwise interactions.
}
\description{
Heatmap overview of cellular networks
}
\details{
A heatmap is generated to represent how much each pair of
cell populations interact including autocrine interactions when
available.

By default, the plot only reports the number of interactions, but it is
possible to provide a list of genes of interest (\code{genes.to.count}
parameter). One option in this
case is to provide receptors involved in specific signaling whose
abundance is to be illustrated on top of the number of interactions. For
this, \code{only.R.in.genes} must be set to TRUE, the default. If not,
then only interactions with both the ligand and the receptor in
\code{genes.to.count} are considered. This may enable more specific
counting. Lastly, it is possible to choose between color-coding the
proportion of interactions with receptor or both receptor and ligand in
\code{genes.to.count}, or to perform an enrichment analysis. In the
last case, the color-code is based on -log10(P-values).
}
\examples{
print("cellNetHeatmap")
if (FALSE) {
    cellNetHeatmap(scsrcn)
}
}
