% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{cluster<-,SCSRDataModel-method}
\alias{cluster<-,SCSRDataModel-method}
\title{clusters setter (internal use only)}
\usage{
\S4method{cluster}{SCSRDataModel}(x) <- value
}
\arguments{
\item{x}{A SCSRDataModel object}

\item{value}{valut to be set for SCSRDataModel}
}
\description{
clusters setter (internal use only)
}
\keyword{internal}
