% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{cluster}
\alias{cluster}
\alias{cluster,SCSRDataModel-method}
\title{clusters accessor}
\usage{
\S4method{cluster}{SCSRDataModel}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
clusters accessor
}
\keyword{internal}
