% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_analysis.R
\name{cluster_analysis}
\alias{cluster_analysis}
\title{Cluster Analysis}
\usage{
cluster_analysis(
  data,
  genes,
  cluster,
  c.names = NULL,
  dif.exp = TRUE,
  s.pval = 10^-2,
  markers = NULL,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or
UMI counts (note : rownames(data)=genes)}

\item{genes}{a character vector of HUGO official gene symbols of length n}

\item{cluster}{a numeric vector of length m}

\item{c.names}{a vector of cluster names}

\item{dif.exp}{a logical (if TRUE, then computes the diferential gene
expression between the clusters using **edgeR**)}

\item{s.pval}{a value, a fixed p-value threshold}

\item{markers}{a table of cell type signature genes}

\item{write}{a logical}

\item{verbose}{a logical}
}
\value{
The function returns a list comprised of a table of differentially
expressed genes, a table of cell types, and a
table of cell cluster types.
}
\description{
Analysis of the differentially expressed genes in the clusters
and their composition by a marker based approach.
}
\examples{
data=matrix(runif(1000,0,1),nrow=5,ncol=200)
rownames(data) = c("A2M","LRP1","AANAT","MTNR1A","ACE")
cluster=c(rep(1,100),rep(2,100))
cluster_analysis(data,rownames(data),cluster,dif.exp=FALSE)
}
