% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering}
\usage{
clustering(
  data,
  n.cluster = 0,
  n = 10,
  method = c("simlr", "kmeans"),
  plot = TRUE,
  pdf = TRUE,
  write = TRUE
)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or UMI counts (note : rownames(data)=genes)}

\item{n.cluster}{a number, an estimation of the ideal number of clusters is computed if equal to 0}

\item{n}{a number, the maximum to consider for an automatic determination of the ideal number of clusters}

\item{method}{"kmeans" or "simlr"}

\item{plot}{a logical}

\item{pdf}{a logical}

\item{write}{a logical}
}
\value{
The function returns a list containing a numeric vector specifying the cluster assignment for each cell,
a 2D t-SNE projection, and the number of cells per cluster.
}
\description{
Identifies the cell clusters, i.e. the cell subpopulations.
}
\examples{
data=matrix(runif(100000,0,1),nrow=500,ncol=200)
clustering(data,n.cluster=2,method="kmeans")
}
