% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{dge.cluster<-,SCSRDataModel-method}
\alias{dge.cluster<-,SCSRDataModel-method}
\title{dge.cluster setter (internal use only)}
\usage{
\S4method{dge.cluster}{SCSRDataModel}(x) <- value
}
\arguments{
\item{x}{A SCSRDataModel object}

\item{value}{valut to be set for SCSRDataModel}
}
\description{
dge.cluster setter (internal use only)
}
\keyword{internal}
