% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{dge.cluster}
\alias{dge.cluster}
\alias{dge.cluster,SCSRDataModel-method}
\title{Model dge.cluster accessor}
\usage{
\S4method{dge.cluster}{SCSRDataModel}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
Model dge.cluster accessor
}
\keyword{internal}
