% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{.classifyByCell}
\alias{.classifyByCell}
\title{Cell classifier}
\usage{
.classifyByCell(data, coordinates, projection, markers, plot, write, verbose)
}
\arguments{
\item{data}{A counts matrix.}

\item{coordinates}{A dataframe with projection coordinates.}

\item{projection}{A character string with the projection method.}

\item{markers}{A dataframe with the markers of celltypes.}

\item{plot}{A logical.}

\item{write}{A logical.}

\item{verbose}{A logical.}
}
\value{
Return a list with each cell assignment cluster and
other informations.
}
\description{
Classifies each cellga using cell type specific markers.
}
\keyword{internal}
