% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{.classifyByCluster}
\alias{.classifyByCluster}
\title{Cluster classification}
\usage{
.classifyByCluster(data, cluster, c.names, markers, write, verbose)
}
\arguments{
\item{data}{A counts matrix}

\item{cluster}{A vector of cluster ids}

\item{c.names}{A vector with unique cluster names}

\item{markers}{a table of cell type signature genes}

\item{write}{a logical}

\item{verbose}{a logical}
}
\description{
Analysis of the clusters composition by a marker based approach.
}
\keyword{internal}
