% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{.clusterClassifying}
\alias{.clusterClassifying}
\title{Cluster classifier}
\usage{
.clusterClassifying(obj, markers, method, most.variables, plot, write, verbose)
}
\arguments{
\item{obj}{an object of type SCSRDataModel}

\item{markers}{a data frame of cell type signature genes}

\item{method}{the method to use for annotation (by cluster or a library)}

\item{most.variables}{a logical}

\item{plot}{a logical (if TRUE, then plots the number of cells
attributed to one cell type, see below)}

\item{write}{a logical}

\item{verbose}{a logical}
}
\value{
A SCSRDataModel with cluster definition
}
\description{
Classifies each cluster using cell type specific markers.
}
\keyword{internal}
