% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{.convertToHuman}
\alias{.convertToHuman}
\title{Transpose To Human Gene Names}
\usage{
.convertToHuman(
  counts,
  dictionnary = data.frame(Gene.name = "A", row.names = "B")
)
}
\arguments{
\item{counts}{A table or matrix of read counts.}

\item{dictionnary}{A dataframe where first column belong to 
organism of study & rownames are the human gene names.}
}
\value{
Return a counts matrix transposed for Human containing a column of 
original gene names
}
\description{
By default, SingleSignalR is designed to work with Homo Sapiens
In order to work with other species, gene names need to be first converted
to Human following an orthology mapping process.
}
\examples{
print('transposeToHuman')
}
\keyword{internal}
