% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrepare.R
\name{.findOrthoGenes}
\alias{.findOrthoGenes}
\title{Orthologs Gene Names}
\usage{
.findOrthoGenes(from_organism, from_values, method)
}
\arguments{
\item{from_organism}{An organism as defined in Ensembl : 
drerio, mmusculus, celegans, dmelanogaster...This is the source organism 
from which you want to convert the gene names to Homo Sapiens.}

\item{from_values}{A vector of gene names from the current species 
studied.}

\item{method}{3 choices are available ('gprofiler','homologene',
'babelgene') gprofiler is set by  default.}
}
\value{
Return a datraframe with 2 columns containing the gene names
for two species.  
First column is the gene name from the source organism 
and the second column corresponds to the  homologous gene name
in  Homo Sapiens.
This function uses orthogene package to query databases
for homologous genes annotation
}
\description{
By default, SingleSignalR is designed to work with Homo 
Sapiens. In order to work with other species, gene names need to be first 
converted to Human following an orthology mapping process.
}
\examples{
print('findOrthoGenes')
}
\keyword{internal}
