% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{.LRscore}
\alias{.LRscore}
\title{Calculation of the LRscore}
\usage{
.LRscore(l, r, s)
}
\arguments{
\item{l}{a value (or a vector) of the mean ligand expression in the
secreting cluster}

\item{r}{a value (or a vector) of the mean receptor expression in the
receiving cluster}

\item{s}{a value for scaling the score (usually the mean of the whole
read count table, the median or another similar value is possible),
must be over}
}
\value{
a value or a vector
}
\description{
Calculation of the LRscore
}
\keyword{internal}
