% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{An example of single-cell RNA-seq data set}
\format{
A data frame with 1629 rows (genes) and 401 (cells)
}
\usage{
data(example_dataset)
}
\description{
A data set containing simulated read counts
}
\keyword{datasets}
