% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_plot_2.R
\name{expression.plot.2}
\alias{expression.plot.2}
\title{Expression Plot 2}
\usage{
expression.plot.2(data, name.1, name.2, tsne)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or UMI counts (note : rownames(data)=genes)}

\item{name.1}{the identifier of the first gene of interest}

\item{name.2}{the identifier of the second gene of interest}

\item{tsne}{a table of n rows and 2 columns with t-SNE projection coordinates for each cell}
}
\value{
The function returns a R plot.
}
\description{
Displays the level of expression of two genes in each cell on the 2D projected data.
}
\examples{
data = matrix(runif(5,0,1),ncol=5)
data[2] = data[5] = 0
rownames(data) = "gene 1"
tsne = matrix(runif(10,0,1),ncol=2)
expression.plot.2(data,"gene 1","gene 2",tsne)

}
