% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_plot.R
\name{expression.plot}
\alias{expression.plot}
\title{Expression Plot}
\usage{
expression.plot(data, name, tsne, colors = c("default", "rainbow", "heat"))
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or UMI counts (note : rownames(data)=genes)}

\item{name}{the identifier of the gene of interest}

\item{tsne}{a table of n rows and 2 columns with 2D projection coordinates for each cell}

\item{colors}{"default" returns the default colorpanel, also accepts "rainbow" or "heat"}
}
\value{
The function returns a R plot.
}
\description{
Displays the level of expression of a gene in each cell on the 2D projected data.
}
\examples{
data = matrix(runif(5,0,1),ncol=5)
data[2] = data[5] = 0
rownames(data) = "gene 1"
tsne = matrix(runif(10,0,1),ncol=2)
expression.plot(data,"gene 1",tsne)

}
