% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_plot_2.R
\name{expression_plot_2}
\alias{expression_plot_2}
\title{Expression Plot 2}
\usage{
expression_plot_2(data, name.1, name.2, tsne)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or UMI counts (note : rownames(data)=genes)}

\item{name.1}{the identifier of the first gene of interest}

\item{name.2}{the identifier of the second gene of interest}

\item{tsne}{a table of n rows and 2 columns with t-SNE projection coordinates for each cell}
}
\value{
The function returns a R plot.
}
\description{
Displays the level of expression of two genes in each cell on the 2D projected data.
}
\details{
This function can be used independantly from any other. It displays
the expression level of two genes of interest on a 2D projection.

`name.1` and `name.2` can be any characters that correspond to a row name of `data`.
}
\examples{
data = matrix(runif(100,0,1),nrow=2,ncol=50)
rownames(data) = c("gene 1", "gene 2")
tsne = matrix(runif(100,-1,1),ncol=2)
expression_plot_2(data,"gene 1","gene 2",tsne)

}
