% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionPlot.R
\name{expressionPlot}
\alias{expressionPlot}
\title{Expression Plot}
\usage{
expressionPlot(
  obj,
  name,
  colors = c("default", "rainbow", "heat"),
  most.variables = TRUE
)
}
\arguments{
\item{obj}{an SCSRDataModel object}

\item{name}{the identifier of the gene of interest}

\item{colors}{'default' returns the default colorpanel, also accepts 
'rainbow' or 'heat'}

\item{most.variables}{a logical}
}
\value{
The function returns a R plot.
}
\description{
Displays the level of expression of a gene in each cell 
on the 2D projected data.
}
\details{
`name` can be any character that corresponds to a row name
of the matrix counts present in the object.

`colors` must be 'default', 'rainbow' or 'heat' exclusively. 'rainbow' and
'heat' are the color palettes provided in R.
}
\examples{
message("expressionplot")
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
#expressionPlot(obj, 'gene 20')
}
