% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionPlot2.R
\name{expressionPlot2}
\alias{expressionPlot2}
\title{Expression Plot 2}
\usage{
expressionPlot2(obj, name.1, name.2, most.variables = TRUE)
}
\arguments{
\item{obj}{an SCSRDataModel object}

\item{name.1}{the identifier of the first gene of interest}

\item{name.2}{the identifier of the second gene of interest}

\item{most.variables}{a logical}
}
\value{
The function returns a R plot.
}
\description{
Displays the level of expression of two genes in each cell 
on the 2D projected data.
}
\details{
This function can be used independantly from any other. It displays
the expression level of two genes of interest on a 2D projection.

`name.1` and `name.2` can be any characters that correspond to a row name 
of the countmatrix present in the object.
}
\examples{
message("expressionplot")
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
#expressionPlot2(obj, 'gene 20', 'gene 2')
}
