% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPathways.R
\name{findPathways}
\alias{findPathways}
\title{find pathways}
\usage{
findPathways(keyword)
}
\arguments{
\item{keyword}{a character string}
}
\value{
The function returns a dataframe containing pathway names that 
contain the specified keyword.
}
\description{
Find pathway names within the database to use in functions.
}
\details{
The `keyword` parameter is set to find pathways that contain it.
}
\examples{
message('findPathways')
df <- findPathways('immune')
}
