% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNet-class.R
\name{infParamSC<-,SCSRNet-method}
\alias{infParamSC<-,SCSRNet-method}
\title{Inference parameters setter (internal use only)}
\usage{
\S4method{infParamSC}{SCSRNet}(x) <- value
}
\arguments{
\item{x}{SCSRNet object.}

\item{value}{value to be set.}
}
\value{
returns \code{NULL}
}
\description{
Inference parameters setter (internal use only)
}
\keyword{internal}
