% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{initialOrganism}
\alias{initialOrganism}
\alias{initialOrganism,SCSRDataModel-method}
\title{organism accessor}
\usage{
\S4method{initialOrganism}{SCSRDataModel}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
organism accessor
}
\keyword{internal}
