% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionDotPlot.R
\name{interactionDotPlot}
\alias{interactionDotPlot}
\title{Interaction Dot Plot}
\usage{
interactionDotPlot(obj, dm = NULL, interaction, pathway = NULL)
}
\arguments{
\item{obj}{an SCSRInference object}

\item{dm}{an SCSRDataModel object}

\item{interaction}{the identifier of the interaction number in the 
interaction dataframe.}

\item{pathway}{the name of a pathway of interest.}
}
\value{
The function returns a R plot.
}
\description{
Displays the level of expression of a LR pair in each cell.
}
\details{
`interaction` must be a number that corresponds to an interaction in the
SCSRInference object.

if `pathway` is set then only receptors involved in those pathways will
be plotted.

if `pathway` is set then the `dm` argument needs to be specified to give
information over the dataset.
}
\examples{
message('interactionDotPlot')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--add Clustering')
obj <- addClustering(obj,cluster.id = sample(1:5,size = 20,replace = TRUE))
obj.int <- cellSignaling(obj)
#interactionDotPlot(obj.int, interaction = 1)
}
