% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interNetwork.R
\name{interNetwork}
\alias{interNetwork}
\title{inter network}
\usage{
interNetwork(
  obj,
  dm,
  clusters = NULL,
  pathway = NULL,
  plot = TRUE,
  most.variables = TRUE,
  write = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{an object of type SCSRInference}

\item{dm}{an object of type SCSRDataModel}

\item{clusters}{a character vector of the clusters names to study}

\item{pathway}{a character vector of pathway names}

\item{plot}{a logical}

\item{most.variables}{a logical}

\item{write}{a logical (if TRUE writes graphML and text files for the
interface and internal networks)}

\item{verbose}{a logical}
}
\value{
The function returns a list containing the tables of interaction
between two cell types and the table for the full network of all the cell
types.
}
\description{
Computes intercellular gene networks.
}
\details{
The `pathway` parameter can be set to select one or several    
specific pathways to study.

If `write` is TRUE, then the function writes four different files. A graphML
file in the *cell-signaling* folder for intercellular interactions between
each pair of clusters named 'intercell_network_Z~1~-Z~2~.graphml', where
Z~1~ and Z~2~ are the *c.names* of the clusters. A graphML file in the
*cell-signaling* folder that contains a compilation of all the 
intercellular, ligand-receptor interactions named 
'full-intercellular-network.graphml'. A text and a graphML file in the 
*networks* folder containing the intracellularnetwork for each cell cluster 
named 'intracell_network_Z.txt' and 'intracell_network_Z.graphml', where Z 
is the *c.names* of the cluster.
}
\examples{
message('interNetwork')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--add Clustering')
obj <- addClustering(obj,cluster.id = sample(1:5,size = 20,replace = TRUE))
message('--cell Signaling')
obj.int <- cellSignaling(obj,int.type = 'paracrine')
#if (!is.null(obj.int)) net <- interNetwork(obj.int, obj)
}
