% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_network.R
\name{intra_network}
\alias{intra_network}
\title{intra network}
\usage{
intra_network(
  goi,
  data,
  genes,
  cluster,
  coi,
  cell.prop = 0.2,
  c.names = NULL,
  signal = NULL,
  write = TRUE,
  plot = TRUE,
  add.lig = TRUE,
  species = c("homo sapiens", "mus musculus"),
  connected = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{goi}{gene of interest (typically a receptor)}

\item{data}{a data frame of n rows (genes) and m columns (cells) of read or
UMI counts (note : rownames(data)=genes)}

\item{genes}{a character vector of HUGO official gene symbols of length n}

\item{cluster}{a numeric vector of length m}

\item{coi}{name of the cluster of interest}

\item{cell.prop}{a threshold, only the genes expressed in this proportion of
the cells of the coi will be taken into account}

\item{c.names}{(optional) cluster names}

\item{signal}{(optional) a list (result of the **cell_signaling()** function)}

\item{write}{a logical (if TRUE writes graphML and text files for the
internal networks)}

\item{plot}{a logical}

\item{add.lig}{a logical (if TRUE adds the goi associated ligands from
signal to the network)}

\item{species}{"homo sapiens" or "mus musculus"}

\item{connected}{a logical (if TRUE keeps only the genes connected to
the goi)}

\item{verbose}{a logical}
}
\value{
The function returns a list containing the internal networks
linked to the genes of interest (goi)
}
\description{
Computes intracellular networks linked to genes of interest.
}
\examples{
data=matrix(runif(1000,0,1),nrow=5,ncol=200)
genes = c("A2M","LRP1","AANAT","MTNR1A","ACE")
cluster=c(rep(1,100),rep(2,100))
intra_network(goi=c("TGFBR1","ERBB2"),data,genes,cluster,coi="T-cells")
}
