% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraNetwork.R
\name{intraNetwork}
\alias{intraNetwork}
\title{intra network}
\usage{
intraNetwork(
  dm,
  goi,
  coi,
  obj = NULL,
  cell.prop = 0.2,
  pathway = NULL,
  write = TRUE,
  plot = TRUE,
  add.lig = TRUE,
  most.variables = TRUE,
  connected = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dm}{a SCSRDataModel object}

\item{goi}{gene of interest (typically a receptor)}

\item{coi}{name of the cluster of interest}

\item{obj}{a SCSRInference object}

\item{cell.prop}{a threshold, only the genes expressed in this proportion of
the cells of the coi will be taken into account}

\item{pathway}{a character vector of pathway names}

\item{write}{a logical (if TRUE writes graphML and text files for the
internal networks)}

\item{plot}{a logical}

\item{add.lig}{a logical (if TRUE adds the goi associated ligands from
signal to the network)}

\item{most.variables}{a logical}

\item{connected}{a logical (if TRUE keeps only the genes connected to
the goi)}

\item{verbose}{a logical}
}
\value{
The function returns a list containing the internal networks
linked to the genes of interest (goi)
}
\description{
Computes intracellular networks linked to genes of interest.
}
\details{
coi must correspond to one of the object's cluster names. If no names was 
given the clusters are numbered such as 'cluster 1', 'cluster 2', ...

`cell.prop` is set to 0.2
by default to avoid unreadable downstream networks. However if the 
calculated network is too small or non-existent (or too big) the user can 
try lower (or higher) values.

The `pathway` parameter can be set to select one or several    
specific pathways to study.

If `most.variables` is TRUE, then the function uses the most 
variable genes matrix counts if it exists in the object.

If `write` is TRUE, then the function writes two different
files. A graphML file in the *network* folder for intracellular
interactions downstream the gene of interest (goi) named
'intracell_network_coi-receptors.graphml'. A text file in the *network*
folder containing the information about the pathways in which the 
interactions are in, named 
'intracell_network_pathway_analysis_coi-receptors.txt'.
}
\examples{
message('intraNetwork')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--add Clustering')
obj <- addClustering(obj,cluster.id = sample(1:5,size = 20,replace = TRUE))
message('--cell Signaling')
obj.int <- cellSignaling(obj,int.type = 'paracrine')
#if (!is.null(obj.int)) net <- intraNetwork(obj, 'gene 20', 1, obj.int)
}
