% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRInference-class.R
\name{ligands}
\alias{ligands}
\alias{ligands,SCSRInference-method}
\title{ligands accessor}
\usage{
\S4method{ligands}{SCSRInference}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
ligands accessor
}
\keyword{internal}
