% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRInference-class.R
\name{LRinter}
\alias{LRinter}
\alias{LRinter,SCSRInference-method}
\title{LRinter accessor}
\usage{
\S4method{LRinter}{SCSRInference}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
LRinter accessor
}
\keyword{internal}
