% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRInference-class.R
\name{LRsubpop}
\alias{LRsubpop}
\alias{LRsubpop,SCSRInference-method}
\title{LRsubpop accessor}
\usage{
\S4method{LRsubpop}{SCSRInference}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
LRsubpop accessor
}
\keyword{internal}
