% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers}
\alias{markers}
\title{Markers}
\usage{
markers(category = c("immune", "tme", "melanoma", "bc"))
}
\arguments{
\item{category}{one or several of the following "immune", "tme", "melanoma",
"bc"}
}
\value{
The function returns a cell type specific markers table.
}
\description{
Provides a table of cell type specific markers.
}
\examples{
markers(c("immune","bc"))

}
