% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers_default.R
\docType{data}
\name{markers_default}
\alias{markers_default}
\title{A list of cell types markers}
\format{
A data frame with 95 rows of 15 variables:
\describe{
  \item{TNBC}{Triple Negative Breast Cancer markers}
  \item{HER+}{HER+ Breast Cancer markers}
  \item{ER+}{ER+ Breast Cancer markers}
  \item{T-cells}{T cells markers}
  \item{B-cells}{B cells markers}
  \item{Macrophages}{Macrophages markers}
  \item{Endothelial cells}{Endothelial cells markers}
  \item{CAFs}{Cancer Associated Fibroblasts markers}
  \item{melanoma}{Melanoma markers}
  \item{Cytotoxic cells}{Cytotoxic cells markers}
  \item{DC}{Dendritic cells markers}
  \item{Mast cells}{Mastocytes cells markers}
  \item{Neutrophils}{Neutrophils cells markers}
  \item{NK cells}{Natural Killer cells markers}
  \item{Treg}{Regulatory T-cells markers}
  ...
}
}
\source{
\url{EDF R\&D}
}
\usage{
markers_default
}
\description{
A list of cell types markers
}
\keyword{datasets}
