% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mouse}
\alias{Mouse}
\title{Example dataset - Mouse}
\format{
A data frame with 500 rows of 425 variables:
}
\source{
\url{EDF R\&D}
}
\usage{
data(Mouse)
}
\description{
Example dataset - Mouse
}
\keyword{dataset}
