% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_interactions.R
\name{mv_interactions}
\alias{mv_interactions}
\title{most variable interactions}
\usage{
mv_interactions(
  data,
  genes,
  cluster,
  c.names = NULL,
  n = 30,
  species = c("homo sapiens", "mus musculus")
)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or
UMI counts (note : rownames(data)=genes)}

\item{genes}{a character vector of HUGO official gene symbols of length n}

\item{cluster}{a numeric vector of length m}

\item{c.names}{(optional) cluster names}

\item{n}{an integer the number of most variables interactions}

\item{species}{"homo sapiens" or "mus musculus"}
}
\value{
The function displays a heatmap showing the most variable
interactions over all clusters
}
\description{
Displays a heatmap showing the most variable interactions over
all clusters.
}
\examples{
data <- matrix(runif(1000,0,1),nrow=5,ncol=200)
genes <- c("gene 1","gene 2","gene 3","gene 4","gene 5")
cluster <- c(rep(1,100),rep(2,100))
mv_interactions(data,genes,cluster)
}
