% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvInteractions.R
\name{mvInteractions}
\alias{mvInteractions}
\title{most variable interactions}
\usage{
mvInteractions(obj, n = 30, most.variables = TRUE, addLR = NULL)
}
\arguments{
\item{obj}{an object of type SCSRDataModel}

\item{n}{an integer the number of most variables interactions}

\item{most.variables}{a logical}

\item{addLR}{a dataframe with 2 columns}
}
\value{
The function displays a heatmap showing the most variable
interactions over all clusters
}
\description{
Displays a heatmap showing the most variable interactions over
all clusters.
}
\details{
The `addLR` allows the user to add LR interaction that they want to study.
It must be a dataframe with one column 'ligand' and one column 'receptor'.
}
\examples{

message('mvInteractions')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--add Clustering')
obj <- addClustering(obj,cluster.id = sample(1:5,size = 20,replace = TRUE))
#mvInteractions(obj)
}
