% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\name{ncounts}
\alias{ncounts}
\alias{ncounts,SCSRDataModel-method}
\title{Counts accessor}
\usage{
\S4method{ncounts}{SCSRDataModel}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
Counts accessor
}
\keyword{internal}
