% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNet-class.R, R/SCSRNoNet-class.R
\name{paracrines}
\alias{paracrines}
\alias{paracrines,SCSRNet-method}
\alias{paracrines,SCSRNoNet-method}
\title{paracrines accessor}
\usage{
\S4method{paracrines}{SCSRNet}(x)

\S4method{paracrines}{SCSRNoNet}(x)
}
\arguments{
\item{x}{SCSRNoNet object}
}
\value{
paracrines

paracrines
}
\description{
paracrines accessor

paracrines accessor
}
\examples{
print("paracrines")
data(example_dataset, package = "SingleCellSignalR")
mat <- log1p(data.matrix(example_dataset[, -1])) / log(2)
rownames(mat) <- example_dataset[[1]]
rme <- rowMeans(mat)
mmat <- mat[rme > 0.05, ]
d <- dist(t(mmat))
h <- hclust(d, method = "ward.D")
pop <- paste0("pop_", cutree(h, 5))
scsrcn <- SCSRNet(mat,
    normalize = FALSE, method = "log-only",
    log.transformed = TRUE, populations = pop
)

paracrines(scsrcn)
}
