% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRInference-class.R
\name{param}
\alias{param}
\alias{param,SCSRInference-method}
\title{Parameters accessor}
\usage{
\S4method{param}{SCSRInference}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
Parameters accessor
}
\keyword{internal}
