% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Patient80}
\alias{Patient80}
\title{Example dataset - Patient80}
\format{
A data frame with 500 rows of 480 variables:
}
\source{
\url{EDF R\&D}
}
\usage{
data(Patient80)
}
\description{
Example dataset - Patient80
}
\keyword{dataset}
