% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw2n.R
\docType{data}
\name{pw2n}
\alias{pw2n}
\title{Number of interactions in pathways}
\format{A data frame with 1703 rows of 2 variables:
\describe{
  \item{pathway}{pathway name}
  \item{N}{Number of interactions}
  ...
}}
\source{
\url{EDF R\&D}
}
\usage{
pw2n
}
\description{
Number of interactions in pathways
}
\keyword{datasets}
