% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PwC_ReactomeKEGG.R
\docType{data}
\name{PwC_ReactomeKEGG}
\alias{PwC_ReactomeKEGG}
\title{Pathway Commons Reactome KEGG 2019-05-08}
\format{
A data frame with 26067 rows of 6 variables:
\describe{
  \item{a.gn}{interactant 1 gene symbol}
  \item{b.gn}{interactant 2 gene symbol}
  \item{type}{interaction type}
  \item{pathway}{Associated pathway}
  ...
}
}
\source{
\url{EDF R\&D}
}
\usage{
PwC_ReactomeKEGG
}
\description{
Pathway Commons Reactome KEGG 2019-05-08
}
\keyword{datasets}
