% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRInference-class.R
\name{receptors}
\alias{receptors}
\alias{receptors,SCSRInference-method}
\title{receptors accessor}
\usage{
\S4method{receptors}{SCSRInference}(x)
}
\arguments{
\item{x}{A SCSRDataModel object}
}
\description{
receptors accessor
}
\keyword{internal}
