% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_signaling.R
\name{score}
\alias{score}
\title{Calculation of the LRscore}
\usage{
score(l, r, s)
}
\arguments{
\item{l}{a value (or a vector) of the mean ligand expression in the
secreting cluster}

\item{r}{a value (or a vector) of the mean receptor expression in the
receiving cluster}

\item{s}{a value for scaling the score (usually the mean of the whole
read count table, the median or another similar value is possible),
must be over 0}
}
\value{
a value or a vector
}
\description{
Calculation of the LRscore
}
\examples{
l=1
r=9
s=5
score(l,r,s)
}
