% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRDataModel-class.R
\docType{class}
\name{SCSRDataModel-class}
\alias{SCSRDataModel-class}
\title{SingleCellSignalR Data Model Object}
\description{
An S4 class to represent expression data used for inferring
ligand-receptor interactions.
}
\section{Slots}{

\describe{
\item{\code{initial.organism}}{Organism}

\item{\code{ncounts}}{List containing read count matrix (row names must be
set to HUGO official gene symbols), most variables count matrix if    
only most variable genes are kept, vector with genes and parameters   
used for cleaning.}

\item{\code{cluster}}{List containing cluster names, number of cells per
cluster and markers used to identify each cluster.}

\item{\code{cell.representation}}{List containing tsne coordinates, cluster
ids and clustering method.}

\item{\code{dge.cluster}}{List containing list of differentially expressed
genes for each cluster and the parameters used for dge.}
}}

\examples{
new('SCSRDataModel', ncounts = list(matrix=matrix(1, nrow=2, ncol=2,
                                   dimnames=list(c('A','B'), c('C','D'))),
                    genes = c('A','B'),
                    param = list(normalization = TRUE, outliers = c(0,0))),
                    initial.organism = 'hsapiens')

}
