% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRInference-class.R
\docType{class}
\name{SCSRInference-class}
\alias{SCSRInference-class}
\title{SingleCellSignalR Interaction Object}
\description{
An S4 class to represent ligand-receptor interactions.
}
\details{
This class is a container for LR interactions along with
their statistical confidence.
}
\section{Slots}{

\describe{
\item{\code{LRinter}}{A list of data frames describing the (ligand,receptor)
doubles with interaction type and LR score for each subpopulation 
combination.}

\item{\code{LRsubpop}}{A data frame describing the number of interactions
for each subpopulation combination.}

\item{\code{ligands}}{A vector of ligands, one entry per LR interaction.}

\item{\code{receptors}}{A vector of receptors, one entry per LR interaction.}

\item{\code{param}}{A list containing the calculation details of interaction.}
}}

\examples{
new('SCSRInference')

}
