% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_network.R
\name{simplify_interactions}
\alias{simplify_interactions}
\title{simplify_interactions}
\usage{
simplify_interactions(t, lr = NULL, autocrine = FALSE)
}
\arguments{
\item{t}{the network to be simplified}

\item{lr}{ligand receptor interactions}

\item{autocrine}{a logical}
}
\value{
t
}
\description{
simplify_interactions
}
\examples{
t=data.frame(a.gn=c("CEP63","CEP63"),b.gn=c("MZT2A","DYNC1L2"),
type=c("in-complex-with","in-complex-with"))
simplify_interactions(t)
}
