% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_interactions}
\alias{visualize_interactions}
\title{Visualize interactions}
\usage{
visualize_interactions(
  inter,
  show.in = NULL,
  write.in = NULL,
  write.out = FALSE,
  method = "default",
  limit = 30
)
}
\arguments{
\item{inter}{a list of data frames result of the **cell_signaling()**
function}

\item{show.in}{a vector of which elements of ```inter``` must be shown}

\item{write.in}{a vector of which elements of ```inter``` must be written}

\item{write.out}{a logical}

\item{method}{a string (usually relative to the experiment)}

\item{limit}{a value between 1 and number of interactions}
}
\value{
The function returns images in the plot window of Rstudio and images
in the pdf format in the *images* folder.
}
\description{
Creates chord diagrams from the interactions tables.
}
\examples{
int.1 = matrix(c("gene 1","gene 1", "gene 2", "gene 3"),ncol=2)
colnames(int.1) = c("cluster 1","cluster 2" )
int.2 = matrix(c("gene 1","gene 4","gene 4","gene 2","gene 3","gene 3"),
ncol=2)
colnames(int.2) = c("cluster 1","cluster 3" )
inter = list(int.1,int.2)
names(inter) = c("1-2","1-3")
visualize_interactions(inter)
}
