% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeInteractions.R
\name{visualizeInteractions}
\alias{visualizeInteractions}
\title{Visualize interactions}
\usage{
visualizeInteractions(
  obj,
  dm,
  pathway = NULL,
  show.in = NULL,
  write.in = NULL,
  write.out = FALSE,
  method = "default",
  limit = 30
)
}
\arguments{
\item{obj}{a SCSRInference object}

\item{dm}{a SCSRDataModel object}

\item{pathway}{a character vector of pathway names}

\item{show.in}{a vector of which elements of ```signal``` must be shown}

\item{write.in}{a vector of which elements of ```signal``` must be written}

\item{write.out}{a logical}

\item{method}{a string (usually relative to the experiment)}

\item{limit}{a value between 1 and number of interactions}
}
\value{
The function returns images in the plot window of Rstudio and 
images in the pdf format in the *images* folder.
}
\description{
Creates chord diagrams from the interactions tables.
}
\details{
The `pathway` parameter can be set to select one specific   
pathway to study.

`show.in` gives the elements of `signal` to be displayed in the 
plot window.

`write.in` gives the elements of `signal` to be written as pdf files
in the *images* folder.

If `write.out` is TRUE, then the function writes a pdf file with
a summary of the all the interactions of `signal` as a chord diagram.

`limit` is the maximum number of interactions displayed on one chord
diagram. Raising this limit over 30 may decrease the visibility.
}
\examples{
message('visualizeInteractions')
message('--dataPrepare')
data <- matrix(runif(1000,0,1),nrow=50,ncol=20)
rownames(data) <- paste('gene',seq_len(50))
obj <- dataPrepare(data)
message('--add Clustering')
obj <- addClustering(obj,cluster.id = sample(1:5,size = 20,replace = TRUE))
message('--cell Signaling')
obj.int <- cellSignaling(obj,int.type = 'paracrine')
#if (!is.null(obj.int)) visualizeInteractions(obj.int, obj)
}
