% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNet-class.R
\name{autocrines<-,SCSRNet-method}
\alias{autocrines<-,SCSRNet-method}
\title{autocrines setter (internal use only)}
\usage{
\S4method{autocrines}{SCSRNet}(x) <- value
}
\arguments{
\item{x}{SCSRNet object}

\item{value}{value to be set for autocrines}
}
\value{
returns \code{NULL}
}
\description{
autocrines setter (internal use only)
}
\keyword{internal}
