% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNoNet-class.R
\name{autocrines<-,SCSRNoNet-method}
\alias{autocrines<-,SCSRNoNet-method}
\title{autocrines setter (internal use only)}
\usage{
\S4method{autocrines}{SCSRNoNet}(x) <- value
}
\arguments{
\item{x}{SCSRNoNet object}

\item{value}{value to be set for autocrines}
}
\value{
returns \code{NULL}
}
\description{
autocrines setter (internal use only)
}
\keyword{internal}
