% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNet-class.R, R/SCSRNoNet-class.R
\name{autocrines}
\alias{autocrines}
\alias{autocrines,SCSRNet-method}
\alias{autocrines,SCSRNoNet-method}
\title{autocrines accessor}
\usage{
\S4method{autocrines}{SCSRNet}(x)

\S4method{autocrines}{SCSRNoNet}(x)
}
\arguments{
\item{x}{SCSRNoNet object}
}
\value{
autocrines

autocrines
}
\description{
autocrines accessor

autocrines accessor
}
\examples{
print("autocrines")
data(example_dataset, package = "SingleCellSignalR")
mat <- log1p(data.matrix(example_dataset[, -1])) / log(2)
rownames(mat) <- example_dataset[[1]]
rme <- rowMeans(mat)
mmat <- mat[rme > 0.05, ]
d <- dist(t(mmat))
h <- hclust(d, method = "ward.D")
pop <- paste0("pop_", cutree(h, 5))
scsrcn <- SCSRNet(mat,
    normalize = FALSE, method = "log-only",
    log.transformed = TRUE, populations = pop
)

autocrines(scsrcn)
}
