% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNoNet-class.R
\name{bsrdmComp<-,SCSRNoNet-method}
\alias{bsrdmComp<-,SCSRNoNet-method}
\title{BSRDataModelComp object setter (internal use only)}
\usage{
\S4method{bsrdmComp}{SCSRNoNet}(x) <- value
}
\arguments{
\item{x}{SCSRNoNet object}

\item{value}{value to be set for bsrdm.comp}
}
\value{
returns \code{NULL}
}
\description{
BSRDataModelComp object setter (internal use only)
}
\keyword{internal}
