% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_signaling.R
\name{cell_signaling}
\alias{cell_signaling}
\title{Cell Signaling}
\usage{
cell_signaling(
  data,
  genes,
  cluster,
  int.type = c("paracrine", "autocrine"),
  c.names = NULL,
  s.score = 0.5,
  logFC = log2(1.5),
  species = c("homo sapiens", "mus musculus"),
  tol = 0,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or
UMI counts (note : rownames(data)=genes)}

\item{genes}{a character vector of HUGO official gene symbols of length n}

\item{cluster}{a numeric vector of length m}

\item{int.type}{"autocrine" or "paracrine"}

\item{c.names}{(optional) cluster names}

\item{s.score}{LRscore threshold}

\item{logFC}{a number, the log fold-change threshold for differentially
expressed genes}

\item{species}{"homo sapiens" or "mus musculus"}

\item{tol}{a tolerance parameter for balancing between "autocrine" and
"paracrine" interactions}

\item{write}{a logical}

\item{verbose}{a logical}
}
\value{
The function returns "autocrine" or "paracrine" interaction lists.
}
\description{
Computes "autocrine" or "paracrine" interactions between cell
clusters.
}
\examples{
data=matrix(runif(1000,0,1),nrow=5,ncol=200)
rownames(data) = c("A2M","LRP1","AANAT","MTNR1A","ACE")
cluster=c(rep(1,100),rep(2,100))
cell_signaling(data,rownames(data),cluster,int.type="paracrine",write=FALSE)
}
