% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{data_prepare}
\alias{data_prepare}
\title{Data Prepare}
\usage{
data_prepare(
  file,
  most.variables = 0,
  lower = 0,
  upper = 0,
  normalize = TRUE,
  write = TRUE,
  verbose = TRUE,
  plot = FALSE
)
}
\arguments{
\item{file}{a string for the scRNAseq data file}

\item{most.variables}{a number}

\item{lower}{a number in [0,1], low quantile threshold}

\item{upper}{a number in [0,1], high quantile threshold}

\item{normalize}{a logical, if TRUE, then computes 99th percentile normalization}

\item{write}{a logical}

\item{verbose}{a logical}

\item{plot}{a logical}
}
\value{
The function returns a data frame of filtered and/or normalized data with genes as row names. If most.variables is equal to N (>0)
the function returns also a sub matrix of the N most variables genes.
}
\description{
Prepares the data for further analysis
}
\examples{
\donttest{data_prepare("~/scRNAseq_dataset.txt")}
}
