% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example dataset}
\format{
A data frame with 1520 rows of 401 variables:
}
\source{
\url{EDF R\&D}
}
\usage{
data(example_dataset)
}
\description{
Example dataset
}
\keyword{dataset}
