% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dataset.R
\docType{data}
\name{example_dataset}
\alias{example_dataset}
\title{Example dataset}
\format{
A data frame with 1520 rows of 401 variables:
}
\source{
\url{EDF R\&D}
}
\usage{
example_dataset
}
\description{
Example dataset
}
\keyword{datasets}
