% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNoNet-class.R
\name{infParamSC<-,SCSRNoNet-method}
\alias{infParamSC<-,SCSRNoNet-method}
\title{Inference parameters setter (internal use only)}
\usage{
\S4method{infParamSC}{SCSRNoNet}(x) <- value
}
\arguments{
\item{x}{SCSRNoNet object.}

\item{value}{value to be set.}
}
\value{
returns \code{NULL}
}
\description{
Inference parameters setter (internal use only)
}
\keyword{internal}
