% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_network.R
\name{inter_network}
\alias{inter_network}
\title{inter network}
\usage{
inter_network(
  data,
  genes,
  cluster,
  signal,
  c.names = NULL,
  species = c("homo sapiens", "mus musculus"),
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data frame of n rows (genes) and m columns (cells) of read or
UMI counts (note : rownames(data)=genes)}

\item{genes}{a character vector of HUGO official gene symbols of length n}

\item{cluster}{a numeric vector of length m}

\item{signal}{a list (result of the **cell_signaling()** function)}

\item{c.names}{(optional) cluster names}

\item{species}{"homo sapiens" or "mus musculus"}

\item{write}{a logical (if TRUE writes graphML and text files for the
interface and internal networks)}

\item{verbose}{a logical}
}
\value{
The function returns a list containing the intercellular networ).
}
\description{
Computes intercellular gene networks.
}
\examples{
m = data.frame(cell.1=runif(10,0,2),cell.2=runif(10,0,2),cell.3=runif(10,0,2),
cell.4=runif(10,0,2),cell.5=runif(10,0,2),cell.6=runif(10,0,2),cell.7=
runif(10,0,2))
rownames(m) = paste("gene", seq_len(10))
cluster = c(1,1,1,2,3,3,2)
inter_network(m,rownames(m),cluster,signal=NULL)
}
