% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRdb.R
\docType{data}
\name{LRdb}
\alias{LRdb}
\title{Ligand/Receptor interactions data table}
\format{
A data frame with 3251 rows of 13 variables:
\describe{
  \item{ligand}{ligand gene symbol}
  \item{receptor}{receptor gene symbol}
  \item{source}{provenance of the interaction}
  \item{PMIDs}{PubmedID of the publication reporting the interaction}
  ...
}
}
\source{
\url{EDF R\&D}
}
\usage{
LRdb
}
\description{
Ligand/Receptor interactions data table
}
\keyword{datasets}
