% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm2Hs.R
\docType{data}
\name{mm2Hs}
\alias{mm2Hs}
\title{Mus Musculus (mm) to Homo Sapiense (Hs) Orthology table}
\format{
A data frame with 16519 rows of 2 variables:
\describe{
  \item{Mouse gene name}{mm gene symbol}
  \item{Gene name}{Hs gene symbol}
  ...
}
}
\source{
\url{EDF R\&D}
}
\usage{
mm2Hs
}
\description{
Mus Musculus (mm) to Homo Sapiense (Hs) Orthology table
}
\keyword{datasets}
