% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNet-class.R
\name{paracrines<-,SCSRNet-method}
\alias{paracrines<-,SCSRNet-method}
\title{paracrines setter (internal use only)}
\usage{
\S4method{paracrines}{SCSRNet}(x) <- value
}
\arguments{
\item{x}{SCSRNet object}

\item{value}{value to be set for paracrines}
}
\value{
returns \code{NULL}
}
\description{
paracrines setter (internal use only)
}
\keyword{internal}
