% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNoNet-class.R
\name{paracrines<-,SCSRNoNet-method}
\alias{paracrines<-,SCSRNoNet-method}
\title{paracrines setter (internal use only)}
\usage{
\S4method{paracrines}{SCSRNoNet}(x) <- value
}
\arguments{
\item{x}{SCSRNoNet object}

\item{value}{value to be set for paracrines}
}
\value{
returns \code{NULL}
}
\description{
paracrines setter (internal use only)
}
\keyword{internal}
