% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCSRNet-class.R
\name{populations<-,SCSRNet-method}
\alias{populations<-,SCSRNet-method}
\title{populations setter (internal use only)}
\usage{
\S4method{populations}{SCSRNet}(x) <- value
}
\arguments{
\item{x}{SCSRNet object}

\item{value}{value to be set for populations}
}
\value{
returns \code{NULL}
}
\description{
populations setter (internal use only)
}
\keyword{internal}
